# backend/api/openapi/v1.yaml
...
openapi: "3.1.0"
info:
  title: VEDO EduTrack API
  description: |
    Educational trajectory service API.
    EduTrack reads knowledge ontologies from VEDO Hub and provides
    route planning, plan execution, gap diagnosis, and visualization.
  version: "0.3.0"
  contact:
    name: VEDO EduTrack Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.edutrack.example.com/api/v1
    description: Production server (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /webhooks/subscriptions:
    get:
      tags: [Webhooks]
      operationId: listWebhookSubscriptions
      summary: List webhook subscriptions
      description: Lists the authenticated tenant's webhook subscriptions, optionally filtered by active status.
      parameters:
        - name: active
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by active status.
      responses:
        "200":
          description: Subscription list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookSubscription"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Webhooks]
      operationId: createWebhookSubscription
      summary: Create a webhook subscription
      description: Creates a subscription for the authenticated tenant and enqueues a verification ping.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSubscriptionCreate"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
        "400":
          description: Invalid subscription payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded (max 5 subscriptions per hour per tenant)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /webhooks/subscriptions/{id}:
    get:
      tags: [Webhooks]
      operationId: getWebhookSubscription
      summary: Get a webhook subscription
      description: Returns one subscription by id (tenant-scoped).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription id.
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Webhooks]
      operationId: updateWebhookSubscription
      summary: Update a webhook subscription
      description: Updates the subscription URL, event types, or signing secret; resets the failure counter.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSubscriptionUpdate"
      responses:
        "200":
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
        "400":
          description: Invalid subscription payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Webhooks]
      operationId: deleteWebhookSubscription
      summary: Delete a webhook subscription
      description: Soft-deletes a subscription; no further deliveries are attempted.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subscription id.
      responses:
        "204":
          description: Subscription deleted
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Subscription not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT signed with RS256, verified against the local JWKS.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error detail.
        endpoint:
          type: string
          description: Endpoint that produced the error (for 501 stubs).
...
    WebhookSubscription:
      type: object
      required: [id, tenant_id, url, event_types, active]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
        url:
          type: string
          format: uri
          description: Subscriber endpoint URL (https; http only for localhost).
        event_types:
          type: array
          description: Subscribed event types (module.mastered, plan.deviated, route.recalculated, standard.risk_detected).
          items:
            type: string
            enum: [module.mastered, plan.deviated, route.recalculated, standard.risk_detected]
        active:
          type: boolean
        failures:
          type: integer
          description: Consecutive delivery failures (>= 5 deactivates the subscription).
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WebhookSubscriptionCreate:
      type: object
      required: [url, event_types]
      properties:
        url:
          type: string
          format: uri
          description: Subscriber endpoint URL (https; http only for localhost).
        event_types:
          type: array
          items:
            type: string
            enum: [module.mastered, plan.deviated, route.recalculated, standard.risk_detected]
        secret:
          type: string
          description: Optional signing secret (>= 32 chars); a random one is generated when omitted.
    WebhookSubscriptionUpdate:
      type: object
      properties:
        url:
          type: string
          format: uri
        event_types:
          type: array
          items:
            type: string
            enum: [module.mastered, plan.deviated, route.recalculated, standard.risk_detected]
        secret:
          type: string
          description: Optional new signing secret (>= 32 chars).
...